Implement Dropdown in Query Templates Dropdown
Had to remove some of the unused varmoji markup and styles from the DE in order for the dropdown menu to position correctly. Apparently some of the 3D transform related styles trash fixed positioning in an unpredictable way
